+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 -   Versioning of cached installers now works correctly
+-   Use of ENTER in authentication and game search possible
 
 ## [0.3.0] - 2020-02-29